Add device calendar support and refactor handlers
Introduce device calendar UI and logic: add CalendarDeviceEventsHandler and CalendarDeviceSection for loading and rendering device calendar events, including permission handling and app-settings link. Add SettingsDeviceCalendarComponents and SettingsDeviceCalendarHandler to manage device calendar toggles, permission requests, and preferences. Refactor CalendarRoute and SettingsRouteSections to remove duplicated device-calendar UI/logic and delegate to the new components/handlers. Update CalendarViewModel and SettingsViewModel to use the new handlers. Overall cleanup extracts device-calendar responsibilities into dedicated classes and composables and centralizes preference updates and permission flows.
