+++ android/app/src/main/java/com/daynest/android/feature/settings/SettingsRouteSections.kt
@@ -14,7 +14,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Card
-import androidx.compose.material3.Checkbox
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
@@ -28,6 +27,7 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import com.daynest.android.R
 import com.daynest.android.data.calendar.DeviceCalendar
+import com.daynest.android.ui.ServerUrlPicker
 
 internal fun LazyListScope.settingsServerSection(
     state: SettingsUiState.Content,
@@ -79,17 +79,7 @@ internal fun LazyListScope.settingsPrivacySection(
             },
         )
     }
-    item {
-        SettingToggleCard(
-            title = stringResource(id = R.string.settings_biometric_label),
-            subtitle = stringResource(id = R.string.settings_biometric_hint),
-            checked = state.biometricLockEnabled,
-            onCheckedChange = { onEvent(SettingsUiEvent.UpdateBiometricLockEnabled(it)) },
-        )
-    }
-    item {
-        biometricTimeoutCard(state, onEvent)
-    }
+    settingsBiometricItems(state, onEvent)
     item {
         SettingToggleCard(
             title = stringResource(id = R.string.settings_calendar_sync_label),
@@ -105,62 +95,28 @@ internal fun LazyListScope.settingsPrivacySection(
             },
         )
     }
-    item {
-        SettingToggleCard(
-            title = stringResource(id = R.string.settings_device_calendars_label),
-            subtitle = stringResource(id = R.string.settings_device_calendars_hint),
-            checked = state.showDeviceCalendars,
-            onCheckedChange = { enabled ->
-                handleDeviceCalendarsChanged(
-                    enabled = enabled,
-                    context = context,
-                    permissionLauncher = deviceCalendarPermissionLauncher,
-                    onEvent = onEvent,
-                )
-            },
-        )
-    }
-    if (state.showDeviceCalendars) {
-        if (state.deviceCalendars.isEmpty()) {
-            item { emptyStateText(R.string.settings_device_calendars_empty) }
-        } else {
-            items(state.deviceCalendars, key = { it.id }) { calendar ->
-                DeviceCalendarToggleRow(
-                    calendar = calendar,
-                    checked = calendar.id in state.enabledDeviceCalendarIds,
-                    onCheckedChange = { onEvent(SettingsUiEvent.UpdateDeviceCalendarEnabled(calendar.id, it)) },
-                )
-            }
-        }
-    }
+    settingsDeviceCalendarsToggleAndList(
+        state = state,
+        context = context,
+        deviceCalendarPermissionLauncher = deviceCalendarPermissionLauncher,
+        onEvent = onEvent,
+    )
 }
 
-@Composable
-private fun DeviceCalendarToggleRow(
-    calendar: DeviceCalendar,
-    checked: Boolean,
-    onCheckedChange: (Boolean) -> Unit,
+private fun LazyListScope.settingsBiometricItems(
+    state: SettingsUiState.Content,
+    onEvent: (SettingsUiEvent) -> Unit,
 ) {
-    Card(modifier = Modifier.fillMaxWidth()) {
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Column(modifier = Modifier.weight(1f)) {
-                Text(text = calendar.name, style = MaterialTheme.typography.bodyMedium)
-                if (!calendar.accountName.isNullOrBlank()) {
-                    Text(
-                        text = calendar.accountName,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.outline,
-                    )
-                }
-            }
-            Checkbox(checked = checked, onCheckedChange = onCheckedChange)
-        }
+    item {
+        SettingToggleCard(
+            title = stringResource(id = R.string.settings_biometric_label),
+            subtitle = stringResource(id = R.string.settings_biometric_hint),
+            checked = state.biometricLockEnabled,
+            onCheckedChange = { onEvent(SettingsUiEvent.UpdateBiometricLockEnabled(it)) },
+        )
+    }
+    item {
+        biometricTimeoutCard(state, onEvent)
     }
 }
 
@@ -315,7 +271,7 @@ private fun loadErrorText(
 }
 
 @Composable
-private fun emptyStateText(messageRes: Int) {
+internal fun emptyStateText(messageRes: Int) {
     Text(
         text = stringResource(id = messageRes),
         style = MaterialTheme.typography.bodyMedium,
@@ -341,28 +297,6 @@ private fun handleNotificationsChanged(
         notificationsPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
     } else {
         onEvent(SettingsUiEvent.UpdatePushNotificationsEnabled(true))
-    }
-}
-
-private fun handleDeviceCalendarsChanged(
-    enabled: Boolean,
-    context: Context,
-    permissionLauncher: ActivityResultLauncher<String>,
-    onEvent: (SettingsUiEvent) -> Unit,
-) {
-    if (!enabled) {
-        onEvent(SettingsUiEvent.UpdateShowDeviceCalendars(false))
-        return
-    }
-    if (
-        ContextCompat.checkSelfPermission(
-            context,
-            Manifest.permission.READ_CALENDAR,
-        ) == PackageManager.PERMISSION_GRANTED
-    ) {
-        onEvent(SettingsUiEvent.UpdateShowDeviceCalendars(true))
-    } else {
-        permissionLauncher.launch(Manifest.permission.READ_CALENDAR)
     }
 }
